feat: [refractionPOINT/tracking#3025] Add new "limacharlie users invite" action for inviting a user (#155)
* feat: [refractionPOINT/tracking#3025] Add new "limacharlie users invite
--email=<email address>" command for inviting a user to LimaCharlie.

Keep in mind that this operation operates on the user behalf and needs
to use user scoped API key.

* chore: [refractionPOINT/tracking#3025] Add WIP stub for integration
tests.

* Revert "chore: [refractionPOINT/tracking#3025] Add WIP stub for integration"

This reverts commit 08308856c7d1db32fddfca47284efa55ed9c1b1e.

We decided we will add those tests as part of separate PR since it
requires more work.

* feat: [refractionPOINT/tracking#3025] Add support for pass multiple
emails to "users invite" command and as such, inviting multiple users at
the same time.

Right now the command support comma delimited --email= option or reading
from a file where each email is separated with a new line.

* feat: [refractionPOINT/tracking#3025] Add support for global "--debug"
flag to the CLI.

When this flag is set, traceback will be printed on any exceptions which
are thrown inside by CLI actions.

* fix: [refractionPOINT/tracking#3025] Fix inadverted regression with CLI
argument parsing.

Also print better error on invalid action.

* feat: [refractionPOINT/tracking#3025] Add very basic sanity tests for
most of the CLI commands.

* feat: [refractionPOINT/tracking#3025] Add basic mock based unit tests
for the new users invite command.

* feat: [refractionPOINT/tracking#3025] Add whoami alias for who command.

* feat: [refractionPOINT/tracking#3025] Add changelog file.

* feat: [refractionPOINT/tracking#3025] Instantiate Manager + User class
after checking arguments.

This way errors are printed before manager performs credentials check.

* feat: [refractionPOINT/tracking#3025] Update changelog.

* feat: [refractionPOINT/tracking#3025] Bump version in the preparation for
the next release.

* feat: [refractionPOINT/tracking#3025] Remove unnecessary print statement.

* feat: [refractionPOINT/tracking#3025] Use better variable name.

* fix: [refractionPOINT/tracking#3025] Fix a small typo spotted by
@dzimine-lc.

* feat: [refractionPOINT/tracking#3025] Add basic sanity check to the PR
workflow to verify that the package can be built and installed.

* feat: [refractionPOINT/tracking#3025] Fix small typo in readme.

* feat: [refractionPOINT/tracking#3025] Update tentative release date in
the changelog file.